Change `AddSingleton` to `AddScoped`
Later in the series there is a guide on adding Entity Framewore Core to the app.
When we register the DbContext the app will fail to run. The app fails to run
because DbContext is scoped by default.

If the Service & Repository are registered as Singleton then each time we load the
code it will keep the same instance, whereas the DbContext requires a new instance.

Due to the 'parent' in this case being a Singleton the process will fail.